City: added last fk check
once Event has been added I'll need to add that to this check

+++ src/main/java/com/mcommings/campaigner/services/locations/CityService.java
@@ -7,6 +7,7 @@ import com.mcommings.campaigner.repositories.IGovernmentRepository;
 import com.mcommings.campaigner.repositories.IWealthRepository;
 import com.mcommings.campaigner.repositories.locations.ICityRepository;
 import com.mcommings.campaigner.repositories.locations.ICountryRepository;
+import com.mcommings.campaigner.repositories.locations.IPlaceRepository;
 import com.mcommings.campaigner.repositories.locations.ISettlementTypeRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +20,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import static com.mcommings.campaigner.enums.ErrorMessage.*;
+import static com.mcommings.campaigner.enums.ForeignKey.FK_CITY;
 
 @Service
 public class CityService implements ICity {
@@ -28,16 +30,18 @@ public class CityService implements ICity {
     private final ICountryRepository countryRepository;
     private final ISettlementTypeRepository settlementTypeRepository;
     private final IGovernmentRepository governmentRepository;
+    private final IPlaceRepository placeRepository;
 
     @Autowired
     public CityService(ICityRepository cityRepository, IWealthRepository wealthRepository,
                        ICountryRepository countryRepository, ISettlementTypeRepository settlementTypeRepository,
-                       IGovernmentRepository governmentRepository) {
+                       IGovernmentRepository governmentRepository, IPlaceRepository placeRepository) {
         this.cityRepository = cityRepository;
         this.wealthRepository = wealthRepository;
         this.countryRepository = countryRepository;
         this.settlementTypeRepository = settlementTypeRepository;
         this.governmentRepository = governmentRepository;
+        this.placeRepository = placeRepository;
     }
 
     @Override
@@ -68,7 +72,9 @@ public class CityService implements ICity {
         if (RepositoryHelper.cannotFindId(cityRepository, cityId)) {
             throw new IllegalArgumentException(DELETE_NOT_FOUND.message);
         }
-        //TODO: check if city.getId() is fk before deleting
+        if (RepositoryHelper.isForeignKey(getReposWhereCityIsAForeignKey(), FK_CITY.columnName, cityId)) {
+            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
+        }
         cityRepository.deleteById(cityId);
     }
 
@@ -89,6 +95,12 @@ public class CityService implements ICity {
         cityToUpdate.setFk_country(city.getFk_country());
         cityToUpdate.setFk_settlement(city.getFk_settlement());
         cityToUpdate.setFk_government(city.getFk_government());
+    }
+
+    //TODO: add Event to this list once it's coded
+    private List<CrudRepository> getReposWhereCityIsAForeignKey() {
+        List<CrudRepository> repositories = new ArrayList<>(Arrays.asList(placeRepository));
+        return repositories;
     }
 
     private boolean hasForeignKeys(City city) {
